Create ec2 agent only if a job is not already assigned to an agent.
When a new agent is created and takes time to auto-register (say, more than 2 mins), the next
agent creation request is sent after 2 minutes. The plugin does not check if the job already has an agent assigned, instead accepts the new request and creates one more agent.
Hence, a job has two elastic agents spinned up. This fix will check if an agent is already assigned and is in pending/running state. If not, the plugin will create an agent.